refactor(content): split monolithic content.ts into focused modules
Extract types (types.ts), shared DOM helpers (utils.ts), and the three
analysis functions into dedicated analyzer modules under analyzers/.
content.ts becomes a thin barrel that re-exports everything, preserving
all existing import paths for consumers.
